# backend/internal/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/admin/auth/send/update/phone:
        post:
            tags:
                - AuthService
            description: 发送手机号验证码
            operationId: AuthService_SendUpdatePhoneCode
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SendUpdatePhoneCodeForm'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /api/admin/auth/update/phone:
        put:
            tags:
                - AuthService
            description: 修改个人中心手机号
            operationId: AuthService_UpdateUserPhone
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdatePhoneForm'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /api/admin/auth/update/pwd:
        put:
            tags:
                - AuthService
            description: 修改个人中心密码
            operationId: AuthService_UpdateUserPwd
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdatePwdForm'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /api/admin/auth/update/userProfile:
        put:
            tags:
                - AuthService
            description: 修改个人中心用户信息
            operationId: AuthService_UpdateUserProfile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserProfileForm'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        SendUpdatePhoneCodeForm:
            type: object
            properties:
                phone:
                    type: string
                    description: 手机号
            description: 发送手机验证码表单
...
        UpdatePhoneForm:
            type: object
            properties:
                phone:
                    type: string
                    description: 手机号
                code:
                    type: string
                    description: 验证码
            description: 修改手机表单
        UpdatePwdForm:
            type: object
            properties:
                oldPwd:
                    type: string
                    description: 原密码
                newPwd:
                    type: string
                    description: 新密码
                confirmPwd:
                    type: string
                    description: 确认密码
            description: 修改密码表单
...
        UserProfileForm:
            type: object
            properties:
                userName:
                    type: string
                    description: 用户名
                nickName:
                    type: string
                    description: 昵称
                avatar:
                    type: string
                    description: 头像URL
                gender:
                    type: integer
                    description: 性别
                    format: int32
                phone:
                    type: string
                    description: 手机号
                roleName:
                    type: string
                    description: 角色名
                deptName:
                    type: string
                    description: 部门名
                createdAt:
                    type: string
                    description: 创建时间
...